suricata/requirements.txt
fastapi==0.99.1
uvicorn[standard]==0.24.0
aiohttp==3.8.4
python-dotenv==1.0.1
aiofiles==23.1.0
